Cleanup and formatting.

+++ src/Hagl.cpp
@@ -15,7 +15,7 @@ const std::vector<hagl::Vertex> vertices = {
 int main(int argc, char* argv[])
 {
 	hagl::WindowSystem windowSystem(1024, 768);
-	hagl::RenderSystem renderSystem(windowSystem, vertices.size());
+	hagl::RenderSystem renderSystem(windowSystem, (uint32_t) vertices.size());
 
 	while (true) {
 		windowSystem.handle_events();
